+++ src/index.jsx
@@ -3,13 +3,15 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import './style.scss';
 import {
-  BrowserRouter, Routes, Route,
+  BrowserRouter, Routes, Route, Navigate,
 } from 'react-router';
 
 import Home from './components/Home';
 import Services from './components/Services';
 import PlasterWashers from './components/PlasterWashers';
 import About from './components/About';
+import PlasterWasherCheckout from './components/PlasterWasherCheckout';
+import Shop from './components/Shop';
 
 function FallBack(props) {
   return <div>URL Not Found</div>;
@@ -24,6 +26,9 @@ function App(props) {
           <Route path="/services" element={<Services />} />
           <Route path="/about" element={<About />} />
           <Route path="/plaster-washers" element={<PlasterWashers />} />
+          <Route path="/plaster-washer-checkout" element={<PlasterWasherCheckout />} />
+          <Route path="/cancel" element={<Navigate to="/plaster-washer-checkout" />} />
+          <Route path="/shop" element={<Shop />} />
           <Route path="*" element={<FallBack />} />
         </Routes>
       </div>
