+++ todo.md
@@ -4,3 +4,5 @@
 -add images to services tabs
 -reformat text on services page
 -pick a new font 
+-clean up footer information
+-make sure people know inventory is one day behind online!! call to make sure or just come in :)
